feat: add global stop key and per partiion stop property
Add support for a global stop key. Create an ssm string parameter of the form `/migrate-block-index/${Config.STAGE}/stop` with any value and it will abort all current invocations for that stage (e.g. prod, staging)

Store a `progress` object per partition with the lastEvaluated object and the recordCount so far, to make it easy to see the total progress.

progress also has a stop property which is set to true when the scan is complete to prevent re-scanning the final partition, or restarting the scan if you manually re-invoke. Idempotent!

If you really want to re-run the entire scan, delete all the progess keys from the parameter store first, for a blank slate.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>